refactor: user_id int32 (#11)

+++ module/util/gin.go
@@ -9,7 +9,7 @@ import (
 )
 
 // GetUserIDFromGinContext 从 RequireAuth 处读取 user_id
-func GetUserIDFromGinContext(c *gin.Context) (int64, error) {
+func GetUserIDFromGinContext(c *gin.Context) (int32, error) {
 	userIDstr := c.GetString("user_id")
 	// 未登录
 	if len(userIDstr) == 0 {
@@ -18,7 +18,7 @@ func GetUserIDFromGinContext(c *gin.Context) (int64, error) {
 	// 已登录
 	userID, err := strconv.ParseInt(userIDstr, 10, 64)
 
-	return userID, err
+	return int32(userID), err
 }
 
 // OKWithMsg 返回成功信息
